Refactored home into components
